add delete confirmation and show created time to human readable format.

+++ app/src/main/java/wordtextcounter/details/main/feature/notes/NotesFragment.kt
@@ -4,17 +4,20 @@ import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
 import android.support.v4.app.Fragment
+import android.support.v7.app.AlertDialog
 import android.support.v7.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
+import kotlinx.android.synthetic.main.fragment_input.*
 import kotlinx.android.synthetic.main.fragment_notes.rvNotes
 import wordtextcounter.details.main.R
 import wordtextcounter.details.main.feature.base.BaseFragment
 import wordtextcounter.details.main.feature.base.BaseViewModel
 import wordtextcounter.details.main.feature.notes.NotesViewModel.ViewState
 import wordtextcounter.details.main.store.ReportDatabase
+import wordtextcounter.details.main.util.extensions.showSnackBar
 
 /**
  * A simple [Fragment] subclass.
@@ -60,7 +63,18 @@ class NotesFragment : BaseFragment() {
           viewModel.editReport(it.position)
         }
         is Delete -> {
-          viewModel.deleteReport(it.position)
+          AlertDialog.Builder(context!!)
+              .setTitle(R.string.edit_alert_title)
+              .setMessage(R.string.delete_alert_desc)
+              .setPositiveButton(R.string.yes,
+                  { dialog, _ ->
+                    viewModel.deleteReport(it.position)
+                    dialog.dismiss()
+                  })
+              .setNegativeButton(R.string.no,
+                  { dialog, _ -> dialog.dismiss() })
+              .setIcon(R.drawable.ic_warning_black_24dp)
+              .create().show()
         }
       }
     })
@@ -76,6 +90,18 @@ class NotesFragment : BaseFragment() {
   private fun handleViewState(viewState: ViewState) {
     if (viewState.reports != null) {
       notesAdapter.setReports(viewState.reports)
+    }
+    
+    if (viewState.showError) {
+      if (viewState.errorMessage == null) {
+        activity?.showSnackBar(getString(R.string.generic_error_message))
+      } else {
+        activity?.showSnackBar(viewState.errorMessage)
+      }
+    }
+    
+    if (viewState.successDeletion) {
+      activity?.showSnackBar(getString(R.string.deletion_success))
     }
   }
